added
 - mixins

renamed
 - globals to common

removed
 - unnecessary indigo-pink.css
