+++ demo/index.html
@@ -5,7 +5,7 @@
 	<% if ( htmlWebpackPlugin.options.environment.ENV === "development") { %>
 	<base href="/">
 	<%} else { %>
-	<base href="/ng2-completer/">
+	<base href="/ngx-select-box/">
 	<% } %>
 	<link rel="stylesheet"
 				href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0-alpha.6/css/bootstrap.min.css"
@@ -14,7 +14,6 @@
 	<link href="https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css" rel="stylesheet" integrity="sha384-wvfXpqpZZVQGK6TAh5PVlGOfQNHSoD2xbE+QkPxCAFlNEevoEH3Sl0sibVcOQVnN" crossorigin="anonymous">
 	<link rel="shortcut icon" type="image/x-icon" href="/favicon.ico"/>
 	<link href="https://fonts.googleapis.com/css?family=Open+Sans" rel="stylesheet">
-	<link href="demo/res/css/indigo-pink.css" rel="stylesheet">
 	<style> html, body {
 		font-family: "Open Sans", sans-serif;
 	}</style>
