+++ src/modules/ngx-select/component/ngx-select.component.scss
@@ -1,3 +1,7 @@
+@import "../../../common/mixins";
+
+@include cdk-overlay;
+
 .ngx-select-trigger {
     display: inline-table;
     cursor: pointer;
@@ -113,46 +117,4 @@
 
 ngx-select-option.ngx-active .ngx-option-text {
     background-color: lightgray;
-}
-
-/*===================================
-    Styles for Angular CDK overlay
-====================================*/
-.cdk-overlay-container {
-    position: fixed;
-    z-index: 5000;
-    pointer-events: none;
-    top: 0;
-    left: 0;
-    height: 100%;
-    width: 100%;
-
-    .cdk-overlay-backdrop {
-        position: absolute;
-        top: 0;
-        bottom: 0;
-        left: 0;
-        right: 0;
-        z-index: 1000;
-        pointer-events: auto;
-        transition: opacity .4s cubic-bezier(.25, .8, .25, 1);
-        opacity: 0;
-
-        &.cdk-overlay-backdrop-showing {
-            opacity: .48
-        }
-
-        &.cdk-overlay-transparent-backdrop {
-            background: 0 0;
-        }
-    }
-
-    .cdk-overlay-pane {
-        position: absolute;
-        pointer-events: auto;
-        box-sizing: border-box;
-        z-index: 1000;
-        height: 256px;
-        margin: 5px 0;
-    }
 }
