+++ src/modules/ngx-select/component/ngx-select.component.ts
@@ -43,7 +43,7 @@ import { filter } from 'rxjs/operators/filter';
 import { defer } from 'rxjs/observable/defer';
 import { merge } from 'rxjs/observable/merge';
 
-import { isNill, noop } from '../globals';
+import { isNill, noop, parseNumber, compare } from '../../../common/common';
 import { NgxSelectModel } from './ngx-select-model';
 import { NgxOptionSelectionChange, NgxSelectOptionComponent } from './ngx-select-option/ngx-select-option.component';
 import { NgxSelectTemplate } from './ngx-select-template';
@@ -342,7 +342,7 @@ export class NgxSelectComponent implements OnInit, AfterContentInit, ControlValu
 
     private findOptionByValue(value: any): NgxSelectOptionComponent | undefined {
         return this.options.find((option: NgxSelectOptionComponent) => {
-            return !isNill(option.value) && this.compareWith(option.value, value);
+            return !isNill(option.value) && compare(option.value, value);
         });
     }
 
@@ -497,7 +497,7 @@ export class NgxSelectComponent implements OnInit, AfterContentInit, ControlValu
         const selectPanelEl: HTMLElement = this.selectPanel.nativeElement;
         const panelTop = selectPanelEl.scrollTop;
         const panelHeight = selectPanelEl.offsetHeight;
-        const scrollOffset = activeOptionIndex * itemHeight + this.parseInt(getComputedStyle(selectContentEl).paddingTop);
+        const scrollOffset = activeOptionIndex * itemHeight + parseNumber(getComputedStyle(selectContentEl).paddingTop);
 
         if (scrollOffset < panelTop) {
             selectPanelEl.scrollTop = scrollOffset;
@@ -527,8 +527,8 @@ export class NgxSelectComponent implements OnInit, AfterContentInit, ControlValu
     private getOptionHeight(): number {
         const selectContentEl: HTMLElement = this.selectContent.nativeElement;
 
-        const panelContentPaddingTop = this.parseInt(getComputedStyle(selectContentEl).paddingTop);
-        const panelContentPaddingBottom = this.parseInt(getComputedStyle(selectContentEl).paddingBottom);
+        const panelContentPaddingTop = parseNumber(getComputedStyle(selectContentEl).paddingTop);
+        const panelContentPaddingBottom = parseNumber(getComputedStyle(selectContentEl).paddingBottom);
         const panelContentHeight = selectContentEl.clientHeight - panelContentPaddingTop - panelContentPaddingBottom;
 
         return panelContentHeight / this.options.length;
@@ -558,18 +558,4 @@ export class NgxSelectComponent implements OnInit, AfterContentInit, ControlValu
     private rotateIcon(): string {
         return this.rotateArrow && this.isPanelOpen ? this._arrowIconUp : this._arrowIconDown;
     }
-
-    // TODO: move to helper class
-    private parseInt(value: string | null): number {
-        const radix: number = 10;
-
-        if (isNill(value)) {
-            return 0;
-        }
-
-        return parseInt(value as string, radix);
-    }
-
-    // TODO: move to helper class || use lodash.Equal
-    private compareWith = (o1: any, o2: any) => o1 === o2;
 }
