--- src/modules/ngx-select/globals.ts
@@ -1,7 +1,0 @@
-export function noop() {
-    return;
-}
-
-export function isNill(value: any) {
-    return value == null;
-}
